requirements.txt
# Core APIs & Server
fastapi==0.104.1
uvicorn==0.24.0

# Frontend
streamlit==1.28.2

# Data Processing & ML
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.3.2
sentence-transformers==2.7.0
faiss-cpu>=1.8.0
rank-bm25==0.2.2
transformers==4.42.4
torch==2.3.1

# Utilities
pydantic==2.5.2
python-multipart==0.0.6
spacy==3.7.5
beautifulsoup4==4.12.3
requests==2.32.3
pillow==10.4.0
tqdm==4.66.5
